Move NodeContext to contexts
